requirements.txt
aiogram==3.6.0
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
qrcode[pil]==7.4.2
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.26.0

